+++ src/Features/Core/Portable/EmbeddedLanguages/RegularExpressions/RegexEmbeddedLanguageFeatures.cs
@@ -1,5 +1,6 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using Microsoft.CodeAnalysis.CodeStyle;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.DocumentHighlighting;
 using Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices;
@@ -10,11 +11,11 @@ namespace Microsoft.CodeAnalysis.Features.EmbeddedLanguages.RegularExpressions
     internal class RegexEmbeddedLanguageFeatures : RegexEmbeddedLanguage, IEmbeddedLanguageFeatures
     {
         public IDocumentHighlightsService DocumentHighlightsService { get; }
-        public DiagnosticAnalyzer DiagnosticAnalyzer { get; }
+        public AbstractCodeStyleDiagnosticAnalyzer DiagnosticAnalyzer { get; }
 
         public RegexEmbeddedLanguageFeatures(EmbeddedLanguageInfo info) : base(info)
         {
-            DocumentHighlightsService = new RegexDocumentHighlightsService(info);
+            DocumentHighlightsService = new RegexDocumentHighlightsService(this);
             DiagnosticAnalyzer = new RegexDiagnosticAnalyzer(info);
         }
     }
