+++ src/Workspaces/Core/Portable/EmbeddedLanguages/RegularExpressions/LanguageServices/RegexEmbeddedLanguage.cs
@@ -1,17 +1,55 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System.Threading;
+using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Classification.Classifiers;
 using Microsoft.CodeAnalysis.EmbeddedLanguages.LanguageServices;
+using Microsoft.CodeAnalysis.LanguageServices;
+using Microsoft.CodeAnalysis.Shared.Extensions;
 
 namespace Microsoft.CodeAnalysis.EmbeddedLanguages.RegularExpressions.LanguageServices
 {
     internal class RegexEmbeddedLanguage : IEmbeddedLanguage
     {
+        public readonly EmbeddedLanguageInfo Info;
+
         public ISyntaxClassifier Classifier { get; }
 
         public RegexEmbeddedLanguage(EmbeddedLanguageInfo info)
         {
+            this.Info = info;
+
             Classifier = new RegexSyntaxClassifier(info);
+        }
+
+        internal async Task<(RegexTree tree, SyntaxToken token)?> TryGetTreeAndTokenAtPositionAsync(
+            Document document, int position, CancellationToken cancellationToken)
+        {
+            var root = await document.GetSyntaxRootAsync(cancellationToken).ConfigureAwait(false);
+            var token = root.FindToken(position);
+            var syntaxFacts = document.GetLanguageService<ISyntaxFactsService>();
+            if (RegexPatternDetector.IsDefinitelyNotPattern(token, syntaxFacts))
+            {
+                return null;
+            }
+
+            var semanticModel = await document.GetSemanticModelAsync(cancellationToken).ConfigureAwait(false);
+            var detector = RegexPatternDetector.TryGetOrCreate(semanticModel, this.Info);
+            var tree = detector?.TryParseRegexPattern(token, cancellationToken);
+            if (tree == null)
+            {
+                return null;
+            }
+
+            return (tree, token);
+        }
+
+        internal async Task<RegexTree> TryGetTreeAtPositionAsync(
+            Document document, int position, CancellationToken cancellationToken)
+        {
+            var treeAndToken = await TryGetTreeAndTokenAtPositionAsync(
+                document, position, cancellationToken).ConfigureAwait(false);
+            return treeAndToken?.tree;
         }
     }
 }
